congestion monitoring for cubes

+++ src/comm/simple_net.cc
@@ -19,18 +19,18 @@ SimpleNet::get_ctrl_msg()
 {
   boost::array<char, HEADER_LEN > buf;
   boost::system::error_code error;
-  int len_len = sock.read_some(boost::asio::buffer(buf));
+  size_t len_len = sock.read_some(boost::asio::buffer(buf));
 
   assert(len_len == HEADER_LEN);
-  int32_t len = ntohl( *(reinterpret_cast<int32_t*> (buf.data())));
+  u_int32_t len = ntohl( *(reinterpret_cast<u_int32_t*> (buf.data())));
   
   std::vector<char> buf2(len);
   
-  int hb_len = sock.read_some(boost::asio::buffer(buf2));
+  size_t hb_len = sock.read_some(boost::asio::buffer(buf2));
   assert(hb_len == len);
 
   boost::shared_ptr<ControlMessage>  h(new ControlMessage);
-  h->ParseFromArray(&buf2[0], hb_len);
+  h->ParseFromArray(&buf2[0], (int) hb_len);
   return h;
 }
 
@@ -40,16 +40,16 @@ SimpleNet::get_data_msg()
 {
   boost::array<char, HEADER_LEN > buf;
   boost::system::error_code error;
-  int len_len = sock.read_some(boost::asio::buffer(buf));
+  size_t len_len = sock.read_some(boost::asio::buffer(buf));
 
   assert(len_len == HEADER_LEN);
-  int32_t len = ntohl( *(reinterpret_cast<int32_t*> (buf.data())));
+  u_int32_t len = ntohl( *(reinterpret_cast<u_int32_t*> (buf.data())));
   
   std::vector<char> buf2(len);
-  int hb_len = sock.read_some(boost::asio::buffer(buf2));
+  size_t hb_len = sock.read_some(boost::asio::buffer(buf2));
 
   boost::shared_ptr<DataplaneMessage>  h(new DataplaneMessage);
-  h->ParseFromArray(&buf2[0], hb_len);
+  h->ParseFromArray(&buf2[0], (int) hb_len);
   return h;
 }
 
@@ -57,14 +57,14 @@ boost::system::error_code
 SimpleNet::send_msg(google::protobuf::MessageLite& m)
 {
   boost::system::error_code err;
-  int sz = m.ByteSize();
+  size_t sz = (size_t) m.ByteSize();
   u_int32_t len_nbo = htonl (sz);
-  int nbytes = sz + HEADER_LEN;
+  size_t nbytes = sz + HEADER_LEN;
 
   u_int8_t * msg = new u_int8_t[nbytes];
 
   memcpy(msg, &len_nbo, HEADER_LEN);
-  m.SerializeToArray((msg + HEADER_LEN), sz);
+  m.SerializeToArray((msg + HEADER_LEN), (int) sz);
   sock.send(boost::asio::buffer(msg, nbytes), 0, err);
   delete msg;
   if (err) {
